Made GDT constexpr, tried to make IDT constexpr but failed

+++ src/kernel/boot/gdt/global_descriptor_table.cpp
@@ -20,6 +20,6 @@
 namespace kernel::boot::gdt {
 
     void initializeGlobalDescriptorTable() {
-        loadGdtTable((uint32_t) &gdtPointer);
+        loadGdtTable(&gdtPointer);
     }
 }
